requirements.txt
icalendar>=5.0
requests>=2.31
pyairtable>=2.2
google-api-python-client>=2.100
google-auth>=2.23
twilio>=8.10
python-dotenv>=1.0
pytz>=2023.3
